requirements.txt
django==1.10
django-tables2
pyyaml
datetime
gunicorn==19.6.0  
psycopg2==2.6.2
